refactor: reorganize file structure and update import paths for logs and scenarios
